fix velocity

+++ plugins/audio/SynthKick23.h
@@ -267,8 +267,10 @@ public:
         return *this;
     }
 
-    void noteOn(uint8_t note, uint8_t velocity) override
+    void noteOn(uint8_t note, float velocity) override
     {
+        // TODO use velocity
+        
         // Could change the frequency base on the note...
         // Could change the amplitude base on the velocity...
         sampleIndex = 0;
